Added first version of the cordapp plugin
