+++ gradle-plugins/cordapp/README.md
@@ -1,0 +1,10 @@
+# Cordapp Gradle Plugin
+
+## Purpose
+
+To transform any project this plugin is applied to into a cordapp project that generates a cordapp JAR.
+
+## Effects
+
+Will modify the default JAR task to create a CorDapp format JAR instead [see here](https://docs.corda.net/cordapp-build-systems.html) 
+for more information.
