+++ gradle-plugins/settings.gradle
@@ -4,3 +4,4 @@ include 'quasar-utils'
 include 'cordformation'
 include 'cordform-common'
 include 'api-scanner'
+include 'cordapp'
